requirements.txt
fastapi
uvicorn
pymongo
pillow
pytesseract
PyMuPDF
transformers
torch
python-multipart
pydantic-settings